Fix a bug wherein some composite numbers were being emitted.
The smallest was 57877 (31 * 1867), and it occurred because the
iterators were only ever moving forward by p (the iterator's
prime). However, when the difference between the candidate and an
iterator was more than p, the iterator didn't move forward far enough,
and so a composite that should have been eliminated by an iterator
_after_ the busted one didn't get eliminated.

Added tests; that's how the bug was found.

Also moved files around so as to look more like 1.3 contrib
projects. This project layout was shamelessly cargo-culted from
clojure/data.priority-map.